+++ src/commands/crash.js
@@ -5,7 +5,7 @@ const _fatalError = require("../functions/fatalError");
 const _promptForYN = require("../functions/promptForYN");
 const { GLOBAL_NAME } = require("../variables/aboutConsts");
 
-const crash = (...params) => {
+const crash = (...args) => {
   const AVAILABLE_CRASHES = [`Fatal Error (${GLOBAL_NAME})`, "Hang", "Memory Leak"];
 
   console.log(